+++ sqlite3/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Restore support for SQLCipher (thanks to [@davidmartos96](https://github.com/davidmartos96)!).
 - Fix crash when loading IndexedDB databases with blocks past the file length ([#380](https://github.com/simolus3/sqlite3.dart/issues/380)).
+- Don't uninstall update hooks when closing databases if no update hooks
+  were installed by this package.
 
 ## 3.3.2